move components file
